+++ package.json
@@ -24,11 +24,12 @@
     "typescript": "^3.9.7"
   },
   "scripts": {
+    "preln": "rm -rf ~/.s/components/devsapp.cn/fc-domain/dist",
+    "ln": "ln -s $(pwd)/dist ~/.s/components/devsapp.cn/fc-domain/dist",
     "dev": "npx tsc -w -p tsconfig.json",
     "prebuild": "rm -rf node_modules && rm -rf package-lock.json && npm i && rimraf dist",
-    "build": "npm run fix && npm run lint && npm run esbuild",
+    "build": "npm run fix && npm run esbuild",
     "esbuild": "esbuild src/index.ts --bundle --log-level=error --minify --platform=node --format=cjs --target=node10.4 --external:@serverless-devs/core --outfile=dist/index.js",
-    "postbuild": "cp package.json dist/",
     "test": "jest",
     "test:cov": "jest --coverage",
     "lint": "f2elint scan",
