+++ src/interface.ts
@@ -1,5 +1,3 @@
-import { CustomDomainConfig } from './lib/fc/custom-domain';
-
 export interface IInputs {
   props: IProperties;
   project: {
@@ -13,7 +11,37 @@ export interface IInputs {
   path: any;
 }
 
+export interface ICredentials {
+  AccountID: string;
+  AccessKeyID: string;
+  AccessKeySecret: string;
+  SecurityToken?: string;
+}
+
 export interface IProperties {
   region: string;
   customDomain: CustomDomainConfig;
+}
+
+
+interface RouteConfig {
+  path: string;
+  serviceName: string;
+  functionName: string;
+  qualifier?: string;
+  methods?: string[];
+}
+
+interface CertConfig {
+  certName: string;
+  certificate: string;
+  privateKey: string;
+}
+
+export interface CustomDomainConfig {
+  domainName: string;
+  protocol: 'HTTP' | 'HTTP,HTTPS';
+  routeConfigs: RouteConfig[];
+  certConfig?: CertConfig;
+  certId?: string;
 }
