--- src/lib/profile.ts
@@ -1,7 +1,0 @@
-
-export interface ICredentials {
-  AccountID: string;
-  AccessKeyID: string;
-  AccessKeySecret: string;
-  SecurityToken?: string;
-}
