+++ src/utils/logger.ts
@@ -1,0 +1,5 @@
+import { Logger } from '@serverless-devs/core';
+
+const logger = new Logger('FC-DOMAIN');
+
+export default logger;
