+++ src/utils/stdout-formatter.ts
